Add files via upload
TMS-Tourism Management System
Introduction
Welcome to the Tourism Management System project! This system is designed to streamline and enhance the management of tourism operations, providing a robust platform for both tourists and administrators to interact efficiently. The project aims to digitalize the traditional tourism processes, making them more accessible, user-friendly, and efficient.

Overview
The Tourism Management System is a comprehensive application that covers various aspects of tourism management. It allows users to explore and book tours, manage their itineraries, and provide feedback. On the administrative side, it offers tools to manage tours, handle bookings, and monitor customer interactions.

Key Features
User Registration and Authentication: Secure sign-up and login functionalities for users.
Tour Booking and Management: Users can browse, book, and manage their tours with ease.
Admin Dashboard: A centralized dashboard for administrators to manage tours, bookings, and customer data.
Customer Feedback and Reviews: A platform for users to provide feedback and reviews, helping improve the services.
Purpose
The primary goal of this project is to create a user-friendly platform that simplifies the booking and management of tours. By integrating all necessary features into a single system, we aim to enhance the overall travel experience for tourists while providing a powerful tool for administrators to manage their operations effectively.

Technologies Used
Backend: Java with Spring Boot for building a robust and scalable server-side application.
Frontend: HTML, CSS, PHP and JavaScript for a responsive and interactive user interface.
Database: MySQL for secure and reliable data storage.
Server: XAMPP server
Version Control: Git for version control and GitHub for code hosting and collaboration.